Started version 2.0, program unpacks native rs2 sprite archives into raw formats
