+++ src/settings.properties
@@ -1,5 +1,5 @@
 title=Better Sprite Packer
-version=1.47.1
+version=2.0.0
 debug=false
 version_link=https://www.dropbox.com/s/7aup8jaroasxkbg/bsp_version.txt?dl=1
 creator_link=https://www.rune-server.org/runescape-development/rs2-client/tools/626093-better-sprite-packer.html
